database/requirements.txt
# 数据库相关依赖（可选）

# SQLite支持（Python内置，无需安装）

# PostgreSQL支持
psycopg2-binary>=2.9.0

# 数据处理
pandas>=1.3.0
openpyxl>=3.0.0
